requirements.txt
numpy
torch
torchvision
tqdm
pandas
opencv
timm
apex
yaml
neptune-client
jupyter
albumentations
pydicom
pretrainedmodels
sklearn
